ENH: Consistently use strongly typed enums

+++ Modules/Core/Mesh/test/itkVTKPolyDataWriterTest02.cxx
@@ -105,7 +105,7 @@ itkVTKPolyDataWriterTest02(int argc, char * argv[])
   {
     CellType * cellPointer = cellsItr.Value();
 
-    if (cellPointer->GetType() != 1)
+    if (static_cast<int>(cellPointer->GetType()) != 1)
     {
       const unsigned int numberOfPoints = cellPointer->GetNumberOfPoints();
